+++ doc/CHANGELOG.md
@@ -7,6 +7,7 @@ Platform-specific changes are prefixed with the platform name, otherwise the cha
 - Added . & .. directories to filesystems that lack it [AB]
 - Replaced previous implementation of realpath() to remove license from AUTHORS [AB]
 - Enabled hybrid PVR DR/DMA vertex submission in driver + sped up pvr_prim() [FG]
+- Add thread priority boosting system [Paul Cercueil = PC]
 
 ## KallistiOS version 2.1.0
 - Cleaned up generated stubs files on a make clean [Lawrence Sebald == LS]